valorDiario(int) -> valorDiario(double)

+++ T2_POO/ContratoDeJogos/src/dados/Jogo.java
@@ -7,11 +7,11 @@ public class Jogo {
     private int ano;
     //no T1 era valor minuto, passou a ser valor diario no novo diagrama (T2)
     //private double valorMinuto;
-    private int valorDiario;
+    private double valorDiario;
     private Categoria categoria;
     private Contrato contrato; // veio apos a 8
 
-    public Jogo(int codigo, String nome, int ano, int valorDiario) {
+    public Jogo(int codigo, String nome, int ano, double valorDiario) {
         this.codigo = codigo;
         this.nome = nome;
         this.ano = ano;
@@ -38,7 +38,7 @@ public class Jogo {
         this.codigo = codigo;
     }
 
-    public int getValorDiario() {
+    public double getValorDiario() {
         return this.valorDiario;
     }
